Pathfinding fix, kinda
Edited rooms to have pathfinding by increasing base pathfinding to an insane size, because the pathfinding script currently only sees the stuff in at start. i.e. it doesn't see the walls of the other rooms, so enemies still phase, but now it at least sees the floor, so they target the player properly